+++ MatrixSDK/Contrib/Swift/MXSession.swift
@@ -54,28 +54,15 @@ public extension MXSession {
      Perform an events stream catchup in background (by keeping user offline).
      
      - parameters:
-        - timeout: the max time to perform the catchup
+        - timeout: the max time to perform the catchup (in seconds)
+        - ignoreSessionState: set true to force the sync, otherwise considers session state equality to paused. Defaults to false.
         - completion: A block called when the SDK has completed a catchup, or times out.
         - response: Indicates whether the sync was successful.
      */
-    @nonobjc func backgroundSync(withTimeout timeout: TimeInterval, completion: @escaping (_ response: MXResponse<Void>) -> Void) {
+    @nonobjc func backgroundSync(withTimeout timeout: TimeInterval, ignoreSessionState: Bool = false, completion: @escaping (_ response: MXResponse<Void>) -> Void) {
         let timeoutMilliseconds = UInt32(timeout * 1000)
-        __backgroundSync(timeoutMilliseconds, success: currySuccess(completion), failure: curryFailure(completion))
+        __backgroundSync(timeoutMilliseconds, ignoreSessionState: ignoreSessionState, success: currySuccess(completion), failure: curryFailure(completion))
     }
-    
-    /**
-     Perform an events stream catchup in background (by keeping user offline). This method does not consider the session state.
-     
-     - parameters:
-     - timeout: the max time to perform the catchup
-     - completion: A block called when the SDK has completed a catchup, or times out.
-     - response: Indicates whether the sync was successful.
-     */
-    @nonobjc func initialBackgroundSync(withTimeout timeout: TimeInterval, completion: @escaping (_ response: MXResponse<Void>) -> Void) {
-        let timeoutMilliseconds = UInt32(timeout * 1000)
-        __initialBackgroundSync(timeoutMilliseconds, success: currySuccess(completion), failure: curryFailure(completion))
-    }
-