+++ MatrixSDK/MXSession.m
@@ -694,10 +694,16 @@ typedef void (^MXOnResumeDone)(void);
 
 - (void)backgroundSync:(unsigned int)timeout success:(MXOnBackgroundSyncDone)backgroundSyncDone failure:(MXOnBackgroundSyncFail)backgroundSyncfails
 {
+    //  background sync considering session state
+    [self backgroundSync:timeout ignoreSessionState:NO success:backgroundSyncDone failure:backgroundSyncfails];
+}
+
+- (void)backgroundSync:(unsigned int)timeout ignoreSessionState:(BOOL)ignoreSessionState success:(MXOnBackgroundSyncDone)backgroundSyncDone failure:(MXOnBackgroundSyncFail)backgroundSyncfails
+{
     // Check whether no request is already in progress
     if (!eventStreamRequest)
     {
-        if (MXSessionStatePaused != _state)
+        if (!ignoreSessionState && MXSessionStatePaused != _state)
         {
             NSLog(@"[MXSession] background Sync cannot be done in the current state %tu", _state);
             dispatch_async(dispatch_get_main_queue(), ^{
@@ -715,22 +721,6 @@ typedef void (^MXOnResumeDone)(void);
 
             [self serverSyncWithServerTimeout:0 success:nil failure:nil clientTimeout:timeout setPresence:@"offline"];
         }
-    }
-}
-
-- (void)initialBackgroundSync:(unsigned int)timeout success:(MXOnBackgroundSyncDone)backgroundSyncDone failure:(MXOnBackgroundSyncFail)backgroundSyncfails
-{
-    // Check whether no request is already in progress
-    if (!eventStreamRequest)
-    {
-        NSLog(@"[MXSession] start a initial background Sync");
-        [self setState:MXSessionStateBackgroundSyncInProgress];
-        
-        // BackgroundSync from the latest known token
-        onBackgroundSyncDone = backgroundSyncDone;
-        onBackgroundSyncFail = backgroundSyncfails;
-        
-        [self serverSyncWithServerTimeout:0 success:nil failure:nil clientTimeout:timeout setPresence:@"offline"];
     }
 }